requirements.txt
ipdb==0.13.13
numpy==1.24.3
openai==0.27.7
pandas==2.0.2
Pillow==10.3.0
pyperclip==1.8.2
PyYAML==6.0.1
scipy==1.9.1
six==1.16.0
tiktoken==0.4.0
tqdm==4.65.0
scikit-image==0.20.0